// packages/@xiaoye-react/schedule/src/types.ts
export type DateStringValue = string;

/** DateTime value used by all UI components, format: `YYYY-MM-DD HH:mm:ss` */
export type DateTimeStringValue = string;

/** Day of the week, 0 – Sunday, 1 – Monday, etc. */
export type DayOfWeek = 0 | 1 | 2 | 3 | 4 | 5 | 6;

/** string – dayjs format, callback function – custom formatter */
export type DateLabelFormat = string | ((date: DateStringValue) => string);

/** View level used by Schedule component */
export type ScheduleViewLevel = 'day' | 'week' | 'month' | 'year';

/** Interaction mode used by Schedule components */
export type ScheduleMode = 'static' | 'default';